+++ src/App.js
@@ -1,17 +1,15 @@
+// List of libraries and components to import
 import React, { useState } from "react";
 import { BrowserRouter, Route, Switch } from "react-router-dom";
 import UserContext from "./context/UserContext";
-import AccountManagement from "./components/pages/AccountManagement";
-import NavBar from "./components/layout/NavBar";
-import Home from "./components/pages/home";
-import Data from "./components/pages/Data";
-import Login from "./components/auth/Login";
-import Register from "./components/auth/Register";
-import GetTable from "./components/Table/getTable";
-import Logs from "./components/pages/Logs";
-import profile from "./components/pages/profile";
-
 import "./style.css";
+import AccountManagement from "./components/AccountManagement/AccountManagement";
+import NavBar from "./components/layout/NavBar";
+import Home from "./components/Home/home";
+import Data from "./components/Data/Data";
+import Login from "./components/Login/Login";
+import Logs from "./components/Logs/Logs";
+import profile from "./components/Profile/profile";
 
 export default function App() {
   const [userData, setUserData] = useState({
@@ -31,11 +29,7 @@ export default function App() {
 
                 <Route exact path="/home" component={Home} />
 
-                <Route
-                  exact
-                  path="/accountmanagement"
-                  component={AccountManagement}
-                />
+                <Route exact path="/accountmanagement" component={AccountManagement} />
 
                 <Route exact path="/data" component={Data} />
