+++ src/components/AccountManagement/AccountManagement.js
@@ -1,6 +1,6 @@
 import React from "react";
-import Register from "../AccountManagement/Register";
-import DeleteUsers from "../AccountManagement/DeleteUsers";
+import Register from "./Register/Register";
+import DeleteUsers from "./DeleteUsers/DeleteUsers";
 import { Tabs, Tab, AppBar } from "@material-ui/core";
 import Typography from "@material-ui/core/Typography";
 import Box from "@material-ui/core/Box";
