+++ src/components/layout/NavBar.js
@@ -7,7 +7,6 @@ import Button from "@material-ui/core/Button";
 import IconButton from "@material-ui/core/IconButton";
 import UserContext from "../../context/UserContext";
 import { useHistory } from "react-router-dom";
-import AccountCircle from "@material-ui/icons/AccountCircle";
 
 import { Link } from "react-router-dom";
 
@@ -68,6 +67,7 @@ export default function NavBar() {
                   />
                 </IconButton>
               </Link>
+
               <Link style={navStyle} to="/data">
                 <Button>
                   <Typography className={classes.navbartext}>
@@ -75,14 +75,19 @@ export default function NavBar() {
                   </Typography>
                 </Button>
               </Link>
+
               <Typography className={classes.title}>
                 Account Type: {userData.user.type}
               </Typography>
+
               <Link style={navStyle} to="/home">
                 <Button>
-                  <Typography className={classes.navbartext}>Home</Typography>
+                  <Typography className={classes.navbartext}>
+                    Home
+                  </Typography>
                 </Button>
               </Link>
+
               <Link style={navStyle} to="/accountmanagement">
                 <Button>
                   <Typography className={classes.navbartext}>
@@ -90,11 +95,15 @@ export default function NavBar() {
                   </Typography>
                 </Button>
               </Link>
+
               <Link style={navStyle} to="/logs">
                 <Button>
-                  <Typography className={classes.navbartext}>Logs</Typography>
+                  <Typography className={classes.navbartext}>
+                    Logs
+                  </Typography>
                 </Button>
               </Link>
+
               <Link style={navStyle} to="/profile">
                 <Button>
                   <Typography className={classes.navbartext}>
@@ -102,9 +111,13 @@ export default function NavBar() {
                   </Typography>
                 </Button>
               </Link>
+
               <Button onClick={logout}>
-                <Typography className={classes.navbartext}>Logout</Typography>
+                <Typography className={classes.navbartext}>
+                  Logout
+                </Typography>
               </Button>
+
             </Toolbar>
           </AppBar>
         </div>
@@ -128,17 +141,23 @@ export default function NavBar() {
                   />
                 </IconButton>
               </Link>
+
               <Typography variant="h6" className={classes.title}>
                 Review discrepancies
               </Typography>
+
               <Typography variant="h6" className={classes.title}>
                 Account Type: {userData.user.type}
               </Typography>
+
               <Link style={navStyle} to="/">
                 <Button>
-                  <Typography className={classes.navbartext}>Home</Typography>
+                  <Typography className={classes.navbartext}>
+                    Home
+                  </Typography>
                 </Button>
               </Link>
+
               <Link style={navStyle} to="/profile">
                 {" "}
                 {/* icon that takes the user to their profile page */}
@@ -148,8 +167,11 @@ export default function NavBar() {
                   </Typography>
                 </Button>
               </Link>
+
               <Button onClick={logout}>
-                <Typography className={classes.navbartext}>Logout</Typography>
+                <Typography className={classes.navbartext}>
+                  Logout
+                </Typography>
               </Button>
             </Toolbar>
           </AppBar>
@@ -174,22 +196,31 @@ export default function NavBar() {
                   />
                 </IconButton>
               </Link>
+
               <Typography variant="h6" className={classes.title}>
                 Review discrepancies
               </Typography>
+
               <Typography variant="h6" className={classes.title}>
                 Account Type: {userData.user.type}
               </Typography>
+
               <Link style={navStyle} to="/">
                 <Button>
-                  <Typography className={classes.navbartext}>Home</Typography>
+                  <Typography className={classes.navbartext}>
+                    Home
+                  </Typography>
                 </Button>
               </Link>
+
               <Link style={navStyle} to="/logs">
                 <Button>
-                  <Typography className={classes.navbartext}>Logs</Typography>
+                  <Typography className={classes.navbartext}>
+                    Logs
+                  </Typography>
                 </Button>
               </Link>
+
               <Link style={navStyle} to="/profile">
                 {" "}
                 {/* icon that takes the user to their profile page */}
@@ -199,9 +230,13 @@ export default function NavBar() {
                   </Typography>
                 </Button>
               </Link>
+
               <Button onClick={logout}>
-                <Typography className={classes.navbartext}>Logout</Typography>
+                <Typography className={classes.navbartext}>
+                  Logout
+                </Typography>
               </Button>
+              
             </Toolbar>
           </AppBar>
         </div>
